feat: move feedback query into parent responses query file
